- Created the cryptocuyrrency dao - Refresh Bitshares balances and transactions - Created the bitshares faucet api to implement the faucet transasctions (work in progress) - Import bitshares account in Account manager - Load bitshares account from db

+++ app/src/main/java/cy/agorise/crystalwallet/models/CryptoCoinBalance.java
@@ -38,7 +38,7 @@ public class CryptoCoinBalance {
     private CryptoCoin mCoin;
 
     @ColumnInfo(name = "balance")
-    private int mBalance;
+    private long mBalance;
 
     public long getId() {
         return mId;
@@ -64,11 +64,11 @@ public class CryptoCoinBalance {
         this.mCoin = coin;
     }
 
-    public int getBalance() {
+    public long getBalance() {
         return mBalance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(long balance) {
         this.mBalance = balance;
     }
